lib: move format and formatWithOptions into internal/util/inspect.js
So these can be required without requiring the whole `util.js`.
